refactor(keys): move parseId and id-length constants to @dxos/keys
- Move SPACE_ID_LENGTH, OBJECT_ID_LENGTH, FQ_ID_LENGTH, and parseId from
  packages/sdk/client/src/echo/util.ts to packages/common/keys/src/parse-id.ts
  (with companion test file).
- Drop the corresponding re-exports from @dxos/client/echo.
- Migrate 5 plugin/app call sites to import from @dxos/keys.
- Add @dxos/keys dep to @dxos/todomvc and tasks app.

Also includes (as part of the same audit follow-up):
- 6 call sites migrated to import SpaceId from @dxos/keys directly
  (instead of going through @dxos/client/echo).
- AUDIT.md link fixes after the file move to src/echo/.
