+++ packages/apps/todomvc/package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "@dxos/config": "workspace:*",
     "@dxos/echo": "workspace:*",
+    "@dxos/keys": "workspace:*",
     "@dxos/log": "workspace:*",
     "@dxos/observability": "workspace:*",
     "@dxos/react-client": "workspace:*",
