+++ packages/apps/todomvc/tsconfig.json
@@ -25,6 +25,9 @@
       "path": "../../common/async"
     },
     {
+      "path": "../../common/keys"
+    },
+    {
       "path": "../../common/log"
     },
     {
