+++ packages/core/compute/functions-runtime/src/edge/functions.ts
@@ -8,7 +8,6 @@ import * as Order from 'effect/Order';
 import { type DID } from 'iso-did/types';
 
 import { type Client } from '@dxos/client';
-import { type SpaceId } from '@dxos/client/echo';
 import { createEdgeIdentity } from '@dxos/client/edge';
 import { Operation } from '@dxos/compute';
 import { Context } from '@dxos/context';
@@ -16,7 +15,7 @@ import { Obj } from '@dxos/echo';
 import { EdgeHttpClient } from '@dxos/edge-client';
 import { FUNCTIONS_META_KEY, setUserFunctionIdInMetadata } from '@dxos/functions';
 import { invariant } from '@dxos/invariant';
-import { type PublicKey } from '@dxos/keys';
+import { type PublicKey, type SpaceId } from '@dxos/keys';
 import { log } from '@dxos/log';
 import { type UploadFunctionResponseBody } from '@dxos/protocols';
 import { safeParseJson } from '@dxos/util';
