+++ packages/plugins/plugin-meeting/src/operations/handle-payload.ts
@@ -5,8 +5,9 @@ import * as Effect from 'effect/Effect';
 import { Capability } from '@dxos/app-framework';
 import { Operation } from '@dxos/compute';
 import { DXN } from '@dxos/echo';
+import { parseId } from '@dxos/keys';
 import { ClientCapabilities } from '@dxos/plugin-client';
-import { Filter, Query, parseId } from '@dxos/react-client/echo';
+import { Filter, Query } from '@dxos/react-client/echo';
 import { type Message } from '@dxos/types';
 
 import { Meeting, MeetingCapabilities, MeetingOperation } from '../types';
