+++ packages/plugins/plugin-space/src/components/CreateDialog/CreateObjectPanel.tsx
@@ -7,7 +7,8 @@ import React, { useCallback, useMemo } from 'react';
 import { type Database, Obj } from '@dxos/echo';
 import { type Collection } from '@dxos/echo';
 import { type AnyProperties } from '@dxos/echo/internal';
-import { type Space, type SpaceId } from '@dxos/react-client/echo';
+import { type SpaceId } from '@dxos/keys';
+import { type Space } from '@dxos/react-client/echo';
 import { toLocalizedString, useDefaultValue, useTranslation } from '@dxos/react-ui';
 import { Form, omitId } from '@dxos/react-ui-form';
 import { SearchList, useSearchListResults } from '@dxos/react-ui-search';
