+++ packages/plugins/plugin-thread/src/types/ThreadOperation.ts
@@ -7,11 +7,11 @@
 import * as Schema from 'effect/Schema';
 
 import { Capability } from '@dxos/app-framework';
+import { SpaceSchema } from '@dxos/client/echo';
 import { Operation } from '@dxos/compute';
 import { Database, Key, Obj, Ref } from '@dxos/echo';
 import { Collection } from '@dxos/echo';
 import { Markdown } from '@dxos/plugin-markdown';
-import { SpaceSchema } from '@dxos/react-client/echo';
 import { Actor, AnchoredTo, Channel, Message, Thread } from '@dxos/types';
 
 import { meta } from '#meta';
