+++ packages/ui/react-ui-canvas-compute/src/shapes/Function.tsx
@@ -9,8 +9,9 @@ import { Script } from '@dxos/compute';
 import { Operation } from '@dxos/compute';
 import { AnyOutput, FunctionInput } from '@dxos/conductor';
 import { Ref, getSnapshot, isInstanceOf } from '@dxos/echo/internal';
+import { parseId } from '@dxos/keys';
 import { useClient } from '@dxos/react-client';
-import { Filter, parseId } from '@dxos/react-client/echo';
+import { Filter } from '@dxos/react-client/echo';
 import {
   type ShapeComponentProps,
   type ShapeDef,
